it_jobs_engine/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
python-dotenv==1.0.1
fake-useragent==1.5.1
flask==3.0.2

